-Added the Simulatie class that will end up outputting all required information regarding the simulation. -Added Car movement using mathematical formulas. -- Cars may not surpass eachother! -getParsedVoertuigen is now called by reference so that it can be used to modify the parsedVoertuigen in the simulation. -Added sortVoertuigen function to sort vectors of class Voertuig based on position (highest comes first). -Parsed vectors of class Voertuig are now automatically sorted to avoid conflict with the mathematical formulas

+++ src/Voertuig.h
@@ -9,6 +9,8 @@
 #define VOERTUIG_H
 
 #include <string>
+#include <vector>
+#include <algorithm>
 using namespace std;
 
 /**
@@ -20,7 +22,7 @@ public:
 
 private:
     string baan;
-    int positie;
+    double positie;
     double snelheid;
     double versnelling;
     double gewensteMaxSnelheid;
@@ -33,7 +35,7 @@ private:
 
 public:
     string getNaamBaan();
-    int getPositie();
+    double getPositie();
     double getSnelheid();
     void setSnelheid(double snelheid);
     double getVersnelling();
@@ -46,7 +48,7 @@ public:
     string getType();
     void setType(string newType);
     void setNaamBaan(string nieuw);
-    void setPositie(int nieuw);
+    void setPositie(double nieuw);
     double getMaxVersnelling();
     int getLengte();
